* 7z/7zip/Bundles/ax7z/SolidCache.cpp: Use unsigned types for indicies.                                        Use unsigned format specifier for unsigned types in OutputDebugPrintf().                                        Use 64-bit type for collective size. * 7z/7zip/Bundles/ax7z/SolidCache.h: OutputDebugPrintf(): Use vsprintf() instead of wvsprintf() in order to show 64-bit values.                                      Use unsigned 64-bit type for collective size and unsigned types for single size.                                      Use unsigned types for indicies.                                      Use unsigned format specifier for unsigned types in OutputDebugPrintf().                                      Add SolidCacheMemory::GetFileCache() const for const correctness.                                      Make GetContent() and OutputContent() const. * 7z/7zip/Bundles/ax7z/SolidCacheDisk.cpp: Capitalize SQL keywords.                                            Use unsigned 64-bit type for collective size and unsigned types for single size.                                            Use unsigned types for indicies.                                            Use unsigned format specifier for unsigned types in OutputDebugPrintf().                                            Output debug messages when actual data files are deleted.                                            SolidCacheDisk::InitDB_(): Delete inmature actual data files when upgrading.                                            SolidCacheDisk::CheckDB_(): Fix missing deletion of actual data files.                                            SolidCacheDisk::PurgeUnmarked_(): Get archive index beforehand.                                            SolidCacheDisk::ReduceSize_(): Fix negative result for unsigned type. * 7z/7zip/Bundles/ax7z/bridge.cpp: GetFileExImp_Cached(): Add type cast in order to avoid warnings. * ax7z.txt: Update build instruction, known issues, TODO and history. * Makefile: Modify targets for tiny test program.
git-svn-id: https://yak.myhome.cx/repos/source/ax7z/trunk@758 ca317041-1a13-0410-b3bf-f3d8fb9af60d

+++ 7z/7zip/Bundles/ax7z/bridge.cpp
@@ -511,13 +511,15 @@ int GetFileExImp_Cached(HLOCAL *dest, const char* pOutFile, UINT32 iExtractFileI
             return SPI_NO_MEMORY;
         }
 
-        scCache.GetContent(iExtractFileIndex, *dest, unpackSize);
+		// Due to Susie plugin specification, size should be in 32 bits.
+        scCache.GetContent(iExtractFileIndex, *dest, static_cast<unsigned int>(unpackSize));
     } else {
         fp = fopen(pOutFile, "wb");
         if (fp == NULL) {
             return SPI_FILE_WRITE_ERROR;
         }
-        scCache.OutputContent(iExtractFileIndex, unpackSize, fp);
+		// Due to Susie plugin specification, size should be in 32 bits.
+        scCache.OutputContent(iExtractFileIndex, static_cast<unsigned int>(unpackSize), fp);
     }
 
     if (fp) {
